+++ dataland-community-manager/src/main/kotlin/org/dataland/datalandcommunitymanager/services/NotificationService.kt
@@ -8,12 +8,12 @@ import org.dataland.datalandcommunitymanager.events.ElementaryEventType
 import org.dataland.datalandcommunitymanager.model.companyRoles.CompanyRole
 import org.dataland.datalandcommunitymanager.repositories.ElementaryEventRepository
 import org.dataland.datalandcommunitymanager.repositories.NotificationEventRepository
+import org.dataland.datalandcommunitymanager.utils.NotificationServiceUtils
 import org.dataland.datalandcommunitymanager.utils.readableFrameworkNameMapping
 import org.dataland.datalandmessagequeueutils.cloudevents.CloudEventMessageHandler
 import org.dataland.datalandmessagequeueutils.constants.ExchangeName
 import org.dataland.datalandmessagequeueutils.constants.MessageType
 import org.dataland.datalandmessagequeueutils.constants.RoutingKeyNames
-import org.dataland.datalandmessagequeueutils.messages.InternalEmailMessage
 import org.dataland.datalandmessagequeueutils.messages.TemplateEmailMessage
 import org.slf4j.LoggerFactory
 import org.springframework.beans.factory.annotation.Autowired
@@ -89,7 +89,10 @@ class NotificationService
                                 unprocessedElementaryEvents,
                             )
                         sendEmailMessagesToQueue(notificationEmailType, emailProperties, emailReceivers, correlationId)
-                        sendInternalMessageToQueue(notificationEmailType, emailProperties, correlationId)
+                        NotificationServiceUtils.sendInternalMessageToQueue(
+                            objectMapper, cloudEventMessageHandler,
+                            notificationEmailType, emailProperties, correlationId,
+                        )
                     }
                 }
         }
@@ -219,46 +222,6 @@ class NotificationService
                     RoutingKeyNames.TEMPLATE_EMAIL,
                 )
             }
-        }
-
-        /**
-         * Sends an internal message about the ir email to the queue.
-         */
-        fun sendInternalMessageToQueue(
-            notificationEmailType: NotificationEmailType,
-            emailProperties: Map<String, String?>,
-            correlationId: String,
-        ) {
-            val keyMap =
-                mapOf(
-                    "baseUrl" to "Environment",
-                    "companyId" to "Company (Dataland ID)",
-                    "companyName" to "Company Name",
-                    "year" to "Reporting periods",
-                    "numberOfDays" to "Number of days",
-                    "framework" to "Framework",
-                    "frameworks" to "Frameworks",
-                )
-
-            val internalEmailProperties = (
-                emailProperties.mapKeys { keyMap[it.key] ?: "Unknown key" } +
-                    mapOf("Notification Email Type" to notificationEmailType.toString())
-            )
-
-            val message =
-                InternalEmailMessage(
-                    "Dataland Notification Email has been send",
-                    "An IR Notification Email has been send",
-                    "IR Notification Email has been send",
-                    internalEmailProperties,
-                )
-            cloudEventMessageHandler.buildCEMessageAndSendToQueue(
-                objectMapper.writeValueAsString(message),
-                MessageType.SEND_INTERNAL_EMAIL,
-                correlationId,
-                ExchangeName.SEND_EMAIL,
-                RoutingKeyNames.INTERNAL_EMAIL,
-            )
         }
 
         /**
